refactor: separate Linux E2E evaluators
